Joining and spawning over network proof of concept

+++ core/src/gmbh/norisknofun/scene/game/figures/Artillery.java
@@ -11,8 +11,8 @@ public class Artillery extends Figure {
 
 
 
-    public Artillery( int x, int y, int width, int height){
-        super(x,y,width,height);
+    public Artillery( float x, float y, float width, float height){
+        super(x,y,width,height, );
         img= new Texture(Gdx.files.internal("img/artillerie.png"));
        super.sprite.setRegion(img);
     }
